Move Squadron helper classes
