+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.MessageReceiver/Startup.cs
@@ -38,7 +38,7 @@ namespace GreenEnergyHub.Charges.MessageReceiver
             ConfigureMessaging(builder);
         }
 
-        public virtual void ConfigureMessaging(IFunctionsHostBuilder builder)
+        public virtual void ConfigureMessaging([NotNull] IFunctionsHostBuilder builder)
         {
             builder.Services
                 .AddMessaging()
